question

+++ resources/views/favorite/favorite_button.blade.php
@@ -1,5 +1,5 @@
 @if (Auth::id() != $micropost['user_id'])
-    {{--@if (Auth::micropost()->is_registering($micropost['user_id'])
+    @if (Auth::micropost()->is_registering($micropost->id))
      {!! Form::open(['route' => ['micropost.unregister', $micropost->id], 'method' => 'delete']) !!}
             {!! Form::submit('Unregister', ['class' => "btn btn-danger btn-block"]) !!}
         {!! Form::close() !!}
@@ -7,5 +7,5 @@
         {!! Form::open(['route' => ['micropost.register', $micropost->id]]) !!}
             {!! Form::submit('Register', ['class' => "btn btn-primary btn-block"]) !!}
         {!! Form::close() !!}
-    @endif--}}
+    @endif
 @endif
